requirements.txt
asgiref==3.7.2
Django==4.2.1
django-bootstrap-datepicker-plus==5.0.3
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
psycopg2==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.6.3
